Fix Stage 2 intro to Stage 2 transition
We were mistakenly passing the stage_state to the slideshow instead of the story_state, which is where we update the step_index.

(I'm not sure if we need the callback and the on_step_index_update. I copied that over from Stage 1. If we don't need it here, someone can go ahead and delete that).