updates to slider
Put slider inside of a card so that it follows the same light/dark color schemes as the graphs. Put the comparison_viewer inside of a card as well. Using classes for these two cards, add styles that eliminate the border radius at the bottom of the comparison viewer, and eliminate the border radius at the top of the slider. Add a margin to the bottom of the comparison viewer to separate slightly.

Also, add grab and grabbing cursors for the slider itself, when user hovers/grabs.