Update when viewers are displayed
Instead of tying display of viewers to a specific list of markers, display the viewer after you pass the marker that is the end of the previous bunch.

At some point we may decide to turn some of them off, but for now, I'm inclined to leave everything that the user has seen and then apply focus to a specific grouping.