Viewer and guideline updates
Note: Viewer layers in this commit are only correct for data being read in. They are incorrect for a first pass through. @Carifio24, I don't know if there is time to fix this, but I probably should have tried to make it correct for the first pass, but I didn't know it would behave like this.