Updates to intro slideshow
Set a minimum height for the window contents. It is set with a calculation from the viewport height, so it is not a very elegant solution.
Still can't figure out how to constrain the hubble telescope image in the first slide! Driving me crazy! As discovered with Pat, choosing an image with a smaller starting resolution works, but that can't be the only way. Surely using the grid system provides functional constraints.