Add remaining slideshow content
Using rv.Img in most cases instead of solara.Image because the latter does not expose alt text prop, which we need to be 508 compliant.

For the more complicated image layouts, I also changed rv.Img to solara.HTML per @johnarban's suggestion because there was a weird bug with the rv.Img that was causing a phantom div to cover up other parts of the page.